Commit de Ronan Le Boulch : - Correction de warnings
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mptk/trunk@1849 0f77665a-23fb-0310-bc23-e91fd8efc3e1

+++ src/libmptk/atom.in.cpp
@@ -198,10 +198,10 @@ int MP_Atom_c::read( FILE *fid, const char mode ) {
   }
 
   /* Check the support information */
-  if ( nItem != ( 2 * (unsigned long int)( numChans ) ) ) {
+  if ( nItem != ( 2 * (int)( numChans ) ) ) {
     mp_error_msg( func, "Problem while reading the supports :"
 		  " %lu read, %lu expected.\n",
-		  nItem, 2 * (unsigned long int )( numChans ) );
+		  nItem, 2 * (int )( numChans ) );
     return( 1 );
   }
